+++ sdk/transfer/index.js
@@ -1,6 +1,8 @@
 const transfer = require('./transfer.js');
 
 exports.log = require('./log');
+exports.Rule = require('./rule/rule.js').Rule;
+exports.rule = require('./rule');
 exports.create = transfer.create;
 exports.delete = transfer.delete;
 exports.query = transfer.query;